+++ src/config/config.go
@@ -92,7 +92,7 @@ func GetArch(archs ...string) string {
 	return runtime.GOARCH
 }
 
-// Get list of valid architectures
+// GetSupportedArchitectures returns a list of valid architectures
 func GetSupportedArchitectures() [] string{
 	return []string{"amd64","arm64"}
 }
