+++ src/pkg/bundle/inspect.go
@@ -8,19 +8,15 @@ import (
 	"context"
 
 	"github.com/defenseunicorns/uds-cli/src/config"
-	"github.com/defenseunicorns/uds-cli/src/pkg/utils"
-	zarfUtils "github.com/defenseunicorns/zarf/src/pkg/utils"
+	"github.com/defenseunicorns/zarf/src/pkg/utils"
 )
 
 // Inspect pulls/unpacks a bundle's metadata and shows it
 func (b *Bundler) Inspect() error {
 	ctx := context.TODO()
 
-	// oci source checks
-	validTarballPath := utils.IsValidTarballPath(b.cfg.InspectOpts.Source)
-	if !validTarballPath {
-		b.cfg.InspectOpts.Source = getOciValidatedSource(b.cfg.InspectOpts.Source)
-	}
+	// Check that provided oci source path is valid, and update it if it's missing the full path
+	CheckOCISourcePath(b)
 
 	// create a new provider
 	provider, err := NewBundleProvider(ctx, b.cfg.InspectOpts.Source, b.tmp)
@@ -47,12 +43,12 @@ func (b *Bundler) Inspect() error {
 		}
 	}
 	// read the bundle's metadata into memory
-	if err := zarfUtils.ReadYaml(loaded[config.BundleYAML], &b.bundle); err != nil {
+	if err := utils.ReadYaml(loaded[config.BundleYAML], &b.bundle); err != nil {
 		return err
 	}
 
 	// show the bundle's metadata
-	zarfUtils.ColorPrintYAML(b.bundle, nil, false)
+	utils.ColorPrintYAML(b.bundle, nil, false)
 
 	// TODO: showing package metadata?
 	// TODO: could be cool to have an interactive mode that lets you select a package and show its metadata
